requirements.txt
coinmarketcap==4.1.1
kafka-python==1.3.5
PyYAML==3.11
prometheus_client==0.1.0
python_telegram_bot==9.0.0
redis==2.10.6
